Prevent link-like behaviour of NavigationItemSwitch
