+++ src/components/navigation-item/common.js
@@ -1,0 +1,11 @@
+export const ICON_SIZE = { inner: 'ms', outer: 'm' };
+
+export const ANIMATION_PROPERTIES = {
+  className: 'navigation-item__content-wrapper',
+  component: 'div',
+  transitionName: 'navigation-item--transition',
+  transitionEnter: true,
+  transitionLeave: true,
+  transitionEnterTimeout: 250,
+  transitionLeaveTimeout: 250
+};
