+++ src/components/navigation-item/index.js
@@ -1,0 +1,37 @@
+/*
+ This file is a part of libertysoil.org website
+ Copyright (C) 2017  Loki Education (Social Enterprise)
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+import omit from 'lodash/omit';
+import PropTypes from 'prop-types';
+import React from 'react';
+
+import NavigationItemLink from './link';
+import NavigationItemPlain from './plain';
+
+export default function NavigationItem(props) {
+  if (props.to) {
+    return <NavigationItemLink {...props} />;
+  }
+
+  return <NavigationItemPlain {...omit(props, ['activeClassName'])} />;
+}
+
+NavigationItem.displayName = 'NavigationItem';
+
+NavigationItem.propTypes = {
+  to: PropTypes.string
+};
