+++ src/less/blocks/page.less
@@ -223,6 +223,8 @@
 
   &--type {
     &_main {
+      position: relative;
+
       // @media @fluidTablet, @fluidDesktopNarrow, @fluidDesktopStandard {
       //   width: 180px;
       // }
